Extract PostScript writer to printer_ps.cpp
Printer::OutputPagePostScript() and Printer::FprintAscii85() move to
src/hardware/printer/printer_ps.cpp. The PS branch has too much
member-state coupling (multipage_output, output_handle,
multipage_counter, ascii85_*, default_page_*, GetPixel) to extract as
a free function cleanly; keeping them as Printer methods in a
separate TU is the right trade-off.

OutputPage's PS branch shrinks to a single OutputPagePostScript()
call. printer.cpp loses ~190 lines.

Also introduces printer_internal.h to export find_next_name across
the printer TUs without exposing it to other modules.
