+++ src/hardware/printer/printer_ps.cpp
@@ -1,0 +1,203 @@
+// SPDX-FileCopyrightText:  2026-2026 The DOSBox Staging Team
+// SPDX-FileCopyrightText:  2002-2013 The DOSBox Team
+// SPDX-License-Identifier: GPL-2.0-or-later
+
+#include "printer.h"
+
+#if C_PRINTER
+
+#include <cstdio>
+#include <optional>
+#include <utility>
+
+#include "misc/std_filesystem.h"
+#include "utils/checks.h"
+
+#include "printer_internal.h"
+
+CHECK_NARROWING();
+
+void Printer::OutputPagePostScript()
+{
+	// If multipage mode is continuing, take ownership back from the
+	// member; otherwise open a fresh file.
+	FILE_unique_ptr psfile = std::move(output_handle);
+
+	if (!psfile) {
+		const auto out_path = find_next_name(multipage_output ? "doc"
+		                                                      : "page",
+		                                     ".ps");
+		if (!out_path) {
+			return;
+		}
+
+		psfile = FILE_unique_ptr{fopen(out_path->string().c_str(), "wb")};
+		if (!psfile) {
+			LOG_ERR("PRINTER: Can't open file %s for printer output",
+			        out_path->string().c_str());
+			return;
+		}
+
+		// Print header.
+		fprintf(psfile.get(), "%%!PS-Adobe-3.0\n");
+		fprintf(psfile.get(), "%%%%Pages: (atend)\n");
+		fprintf(psfile.get(),
+		        "%%%%BoundingBox: 0 0 %i %i\n",
+		        static_cast<uint16_t>(default_page_width * 72),
+		        static_cast<uint16_t>(default_page_height * 72));
+		fprintf(psfile.get(), "%%%%Creator: DOSBOX Virtual Printer\n");
+		fprintf(psfile.get(), "%%%%DocumentData: Clean7Bit\n");
+		fprintf(psfile.get(), "%%%%LanguageLevel: 2\n");
+		fprintf(psfile.get(), "%%%%EndComments\n");
+		multipage_counter = 1;
+	}
+
+	fprintf(psfile.get(), "%%%%Page: %i %i\n", multipage_counter,
+	        multipage_counter);
+	fprintf(psfile.get(),
+	        "%i %i scale\n",
+	        static_cast<uint16_t>(default_page_width * 72),
+	        static_cast<uint16_t>(default_page_height * 72));
+	fprintf(psfile.get(),
+	        "%i %i 8 [%i 0 0 -%i 0 %i]\n",
+	        page->w,
+	        page->h,
+	        page->w,
+	        page->h,
+	        page->h);
+	fprintf(psfile.get(), "currentfile\n");
+	fprintf(psfile.get(), "/ASCII85Decode filter\n");
+	fprintf(psfile.get(), "/RunLengthDecode filter\n");
+	fprintf(psfile.get(), "image\n");
+
+	SDL_LockSurface(page);
+
+	uint32_t pix          = 0;
+	const uint32_t numpix = page->h * page->w;
+	ascii85_buffer_pos    = ascii85_cur_col = 0;
+
+	while (pix < numpix) {
+		// Compress data using RLE.
+		if ((pix < numpix - 2) &&
+		    (GetPixel(pix) == GetPixel(pix + 1)) &&
+		    (GetPixel(pix) == GetPixel(pix + 2))) {
+			// Three or more pixels with the same colour: RLE run.
+			uint8_t sameCount = 3;
+			const uint8_t col = GetPixel(pix);
+			while (sameCount < 128 && sameCount + pix < numpix &&
+			       col == GetPixel(pix + sameCount)) {
+				sameCount++;
+			}
+
+			FprintAscii85(psfile.get(), 257 - sameCount);
+			FprintAscii85(psfile.get(), 255 - col);
+
+			pix += sameCount;
+		} else {
+			// Find end of heterogeneous area.
+			uint8_t diffCount = 1;
+			while (diffCount < 128 && diffCount + pix < numpix &&
+			       ((diffCount + pix < numpix - 2) ||
+			        (GetPixel(pix + diffCount) !=
+			         GetPixel(pix + diffCount + 1)) ||
+			        (GetPixel(pix + diffCount) !=
+			         GetPixel(pix + diffCount + 2)))) {
+				diffCount++;
+			}
+
+			FprintAscii85(psfile.get(), diffCount - 1);
+			for (uint8_t i = 0; i < diffCount; i++) {
+				FprintAscii85(psfile.get(), 255 - GetPixel(pix++));
+			}
+		}
+	}
+
+	// Write end-of-data marker for the RLE + ASCII85 filters.
+	FprintAscii85(psfile.get(), 128);
+	FprintAscii85(psfile.get(), 256);
+
+	SDL_UnlockSurface(page);
+
+	fprintf(psfile.get(), "showpage\n");
+
+	if (multipage_output) {
+		multipage_counter++;
+		output_handle = std::move(psfile);
+	} else {
+		fprintf(psfile.get(), "%%%%Pages: 1\n");
+		fprintf(psfile.get(), "%%%%EOF\n");
+		// psfile closes here via FILE_unique_ptr destructor.
+		output_handle.reset();
+	}
+}
+
+void Printer::FprintAscii85(FILE* file, uint16_t byte)
+{
+	if (byte != 256) {
+		if (byte < 256) {
+			ascii85_buffer[ascii85_buffer_pos++] = static_cast<uint8_t>(byte);
+		}
+
+		if (ascii85_buffer_pos == 4 || byte == 257) {
+			uint32_t num = static_cast<uint32_t>(ascii85_buffer[0])
+			                    << 24 |
+			               static_cast<uint32_t>(ascii85_buffer[1])
+			                       << 16 |
+			               static_cast<uint32_t>(ascii85_buffer[2])
+			                       << 8 |
+			               static_cast<uint32_t>(ascii85_buffer[3]);
+
+			// Deal with special case.
+			if (num == 0 && byte != 257) {
+				fprintf(file, "z");
+				if (++ascii85_cur_col >= 79) {
+					ascii85_cur_col = 0;
+					fprintf(file, "\n");
+				}
+			} else {
+				char buffer[5];
+				for (int8_t i = 4; i >= 0; i--) {
+					buffer[i] = static_cast<uint8_t>(
+					        static_cast<uint32_t>(num) %
+					        static_cast<uint32_t>(85));
+					buffer[i] += 33;
+					num /= static_cast<uint32_t>(85);
+				}
+
+				// Make sure a line never starts with a % (which
+				// may be mistaken as start of a comment).
+				if (ascii85_cur_col == 0 && buffer[0] == '%') {
+					fprintf(file, " ");
+				}
+
+				for (int i = 0;
+				     i < ((byte != 257) ? 5 : ascii85_buffer_pos + 1);
+				     i++) {
+					fprintf(file, "%c", buffer[i]);
+					if (++ascii85_cur_col >= 79) {
+						ascii85_cur_col = 0;
+						fprintf(file, "\n");
+					}
+				}
+			}
+
+			ascii85_buffer_pos = 0;
+		}
+
+	} else {
+		// Close ASCII85 string. Pad partial tuple with zero bytes and
+		// emit '~>' end-of-data marker.
+		if (ascii85_buffer_pos > 0) {
+			for (uint8_t i = ascii85_buffer_pos; i < 4; i++) {
+				ascii85_buffer[i] = 0;
+			}
+
+			FprintAscii85(file, 257);
+		}
+
+		fprintf(file, "~");
+		fprintf(file, ">\n");
+	}
+}
+
+#endif // C_PRINTER
